% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax-sysdata.R
\docType{data}
\name{cax_column_definitions}
\alias{cax_column_definitions}
\title{Column names and definitions}
\description{
An internal dataset with column name and definitions. Produced from the
\href{https://www.streamnet.org/data/hli}{CAX HLI Tabular Query} Excel file definition tabs. It is a data frame with name and definition.
}
\examples{
head(rCAX:::cax_column_definitions)
}
\keyword{datasets}
