% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax-sysdata.R
\docType{data}
\name{caxtabs}
\alias{caxtabs}
\title{CAX API table names and table_id}
\description{
An internal dataset output from a call to "ca/tables" per \href{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}{REST API documentation}. It is a data frame with a number of columns. The name and id columns are the ones used internally.
}
\examples{
tab <- rCAX:::caxtabs
head(tab[,c("name", "id")])
}
\keyword{datasets}
