% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_GET.R
\name{check_key}
\alias{check_key}
\title{Get the key from the user system environment variable}
\usage{
check_key(key)
}
\arguments{
\item{key}{the key}
}
\description{
Saved in \code{CAX_KEY} in the \code{.Renviron} file. Alternatively it can be passed in.
}
\references{
This function is modeled off the functions in \url{https://github.com/ropensci/rredlist}
}
