% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax-sysdata.R
\docType{data}
\name{internal_data}
\alias{internal_data}
\title{Internal data used by functions}
\description{
rCAX has a number of internally stored data (in \code{sysdata.rda}) that are
used by the functions. If the user wants to use these, preface the name
with \verb{rCAX:::}. For example, \code{rCAX:::caxesu} shows the list of ESU/DPS names.
}
\details{
Internal data
\itemize{
\item \code{caxsuperpops}: the SuperPopulations table
\item \code{caxtabs}: the API table names and their table_ids. Used to look up table_id
\item \code{caxpops}: the Populations table
\item \code{caxesu}: the list of ESU/DPS names. Useful for filtering.
\item \code{nosa_xport_colnames}, \code{sar_xport_colnames}, \code{pni_xport_colnames}, \code{rpers_xport_colnames}, \code{presmolt_xport_colnames}, \code{juvout_xport_colnames}: the column names that appear in the HLI Tabular Query output so that rCAX \verb{_xport} functions produce the same columns in the same order.
\item \code{cax_column_definitions}: a data frame with column name and definitions. Produced from the HLI Tabular Query Excel files
}
}
\examples{
rCAX:::caxesu
}
\keyword{datasets}
