% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax-package.R
\docType{package}
\name{rCAX-package}
\alias{rCAX-package}
\alias{rCAX}
\title{rCAX}
\description{
This package is an R client for the StreamNet Coordinated Assessments HLI REST API.
}
\section{Authentication}{

The rCAX R package includes a pull key so that you can query the CAX database. Should you
need to use your own key see the vignette on the API functions.
}

\keyword{package}
