% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_citation.R
\name{rcax_citation}
\alias{rcax_citation}
\title{Get the citation CAX version}
\usage{
rcax_citation(...)
}
\value{
API citation as character string
}
\description{
Get the citation CAX version
}
\examples{
rcax_citation()

}
