% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_GET.R
\name{rcax_GET}
\alias{rcax_GET}
\title{Make a GET call to API}
\usage{
rcax_GET(path, key = NULL, query = NULL, ...)
}
\arguments{
\item{path}{what to add after the base api path}

\item{key}{A CAX API key. See details.}

\item{query}{a list of query parameters with their values, e.g. \code{list(param=value)}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Make a GET call to API
}
\details{
API Key: rCAX includes a "read-only" API key. Thus for normal
read-only use, you
do not need an API key. If you need to use your
own key, see the instructions in \code{vignette("api_get", package = "rCAX")}.

User Agent String: \code{rcax_ua()} sets the user agent string that is passed
to the CAX REST API. If the rCAX package code is reused (for another
package), make sure to change the user agent string to indicate that
the queries are from another package.
}
\examples{
\dontrun{
rcax_GET("ca/tables")
}
}
\references{
API docs at \url{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}

This function is modeled off \code{rl_GET()} in \url{https://github.com/ropensci/rredlist}
}
