% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_hli.R
\name{rcax_hli}
\alias{rcax_hli}
\title{Return HLI base tables}
\usage{
rcax_hli(
  hli = c("NOSA", "SAR", "PNI", "PreSmolt", "JuvOut", "RperS"),
  flist = NULL,
  qlist = NULL,
  cols = NULL,
  sortcols = NULL,
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)
}
\arguments{
\item{hli}{The HLI short name}

\item{flist}{A filter for the query. See details and \code{rcax_filter()}}

\item{qlist}{Additional query params. See \code{rcax_table_query()} for details.}

\item{cols}{column names to return. Use cols=NULL if you want all columns. Names are case insensitive.}

\item{sortcols}{The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. \code{c("spawningyear", "popid", "MPG")}. Names are case insensitive.}

\item{type}{whether to return the table ("data.frame") or colnames with definitions ("colnames"). Default is to return the table as a data frame.}

\item{GETargs}{A list of arguments for the \code{rcax_GET()} call. These do not need to be specified by the user unless the user wants to change the default values. \code{table_id} This is the CAX table id. It is looked up using \code{tablename} and \code{rCAX:::caxpops} (an internal dataset). If \code{table_id} is passed in via \code{GETargs}, it wil override the default table id lookup. \code{recordloc} This is the name of the record we want in the list returned from the GET call. \code{key} is the API key.}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
data frame if \code{type="data.frame"} (default) and the colnames if \code{type="colnames"}
}
\description{
Returns the base HLI table which is missing the
additional metadata (such as NMFS_PopID and update dates) found in the XPort tables. The HLI is specified using its short code:
\itemize{
\item NOSA: Natural Origin Spawner Abundance
\item SAR: Smolt to Adult Ratios
\item PNI: Proportionate Natural Influence of supplementation hatcheries
\item RperS: Recruits per Spawner
\item JuvOut: Juvenile Outmigrants
\item PreSmolt: Presmolt Abundance
}

The query will download 1000 records by default; you can change
this by passing in \code{qlist=list(limit=1100)} for example. To make a
filtered query e.g. for one ESU or popid) pass in the \code{flist}
argument. Field names and values are case insensitive, e.g.
\code{POPID}, \code{popid} and \code{PopID} return the same result.
If your filter value is not in the table, the returned table
will be empty. Examples:
\itemize{
\item \code{flist=list(popid=7)} return values for popid 7.
\item \code{flist=list(esu_dps="Salmon, Chinook (Snake River spring/summer-run ESU)")} return values for one ESU. Use \code{rCAX:::caxesu} to see a list of ESU_DPS names.
}
}
\details{
The table_id is set automatically using a saved data frame from a \code{rcax_tables()} call. The table is saved in \code{R/sysdata.rda}.

If you want to see only the column names use
\verb{type="colnames}.
}
\examples{
# return popid of first record
# Note the part after $ is case sensitive
id <- rcax_hli("NOSA", qlist=list(limit=1))$popid

a <- rcax_hli("NOSA", flist=list(popid=id))
a <- a[, c("popid", "spawningyear", "nosaij", "nosaej")]
head(a)

# to print the first column names and definitions
head(rcax_hli("NOSA", type="colnames"))

}
\references{
API docs at \url{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}
}
\seealso{
\code{rcax_hli_xport()}, \code{rcax_table_query()}, \code{rcax_filter()}
}
