% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_nosa_xport.R
\name{rcax_nosa_xport}
\alias{rcax_nosa_xport}
\title{Return NOSA table from CAP Fish HLIs Tabular Query}
\usage{
rcax_nosa_xport(
  tablename = "XPortCA_NOSA_01",
  flist = NULL,
  qlist = NULL,
  cols = NULL,
  sortcols = c("spawningyear", "popid"),
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)
}
\arguments{
\item{tablename}{(character) The CAX name for the table; ignored if \code{GETargs$table_id} passed in.}

\item{flist}{A filter for the query. See details and \code{rcax_filter()}}

\item{qlist}{Additional query params. See \code{rcax_table_query()} for details.}

\item{cols}{column names to return. Use cols=NULL if you want all columns}

\item{sortcols}{The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. \code{c("spawningyear", "popid", "MPG")}}

\item{type}{whether to return the table ("data.frame") or colnames ("colnames"). Default is to return the table as a data frame.}

\item{GETargs}{A list of arguments for the \code{rcax_GET()} call. These do not normally need to be specified by the user. \code{table_id} This is the CAX table id. It is looked up using \code{tablename} and \code{caxpops}. \code{recordloc} This is the name of the record we want in the list returned from the GET call. \code{key} is the API key. See \code{rcax_GET()} for API key details.}
}
\value{
data frame if \code{type="data.frame"} (default) and the colnames if \code{type="colnames"}
}
\description{
Returns NOSA with some additional metadata (such as NMFS_PopID and
update dates). This is the table that one can download from the
\href{https://www.streamnet.org/data/hli/}{CAP Fish HLIs Tabular Query}.

\code{rcax_nosa_xport()} will download 1000 records. You will
want to make a filtered query by passing in values to filter on
via \code{flist}. Keep in mind that the field names
and values are case insensitive. Using \code{POPID}, \code{popid} and \code{PopID}
will have the same effect. Examples:
\itemize{
\item \verb{rcax_nosa_xport(flist=list(popid=7)} return values for popid 7.
\item \verb{rcax_nosa_xport(flist=list(nmfs_popid=6)} return values for NMFS_PopID 6.
\item \verb{rcax_nosa_xport(flist=list(esu_dps="Salmon, Chinook (Snake River spring/summer-run ESU)")} return values for one ESU. Use \code{rCAX:::caxesu} to see a list of ESU_DPS names.
}
}
\details{
The table_id is set automatically using a saved data frame from a \code{rcax_tables()} call. The table is saved in \code{R/sysdata.rda}.

The table has a few columns that do not appear in the Excel files that
one can download from StreamNet: "species", "publish", "num",
"hli", "agency", "esudps", "hli_id". "esudps" is a slight variant on
"esu_dps" and is removed. "num" is a sort flag and is removed.
"publish" will always be Yes but is left in. The other columns are left in. The colnames are resorted into the order found in the downloaded
Excel files with the extra column added to the end.
}
\examples{
a <- rcax_nosa_xport(
         flist=list(nmfs_popid=7),
         cols=c("nmfs_popid", "spawningyear", "tsaej", "nosaej")
         )
head(a)

# First 3 columns
rcax_nosa_xport(qlist=list(limit=3))[,1:3]

# to print the first 5 column names
rcax_nosa_xport(type="colnames")[1:5]

# to print the first 5 column names as they appear
# in Excel files downloaded from https://www.streamnet.org/data/hli/
tab <- rcax_nosa_xport(qlist=list(limit=1))
colnames(tab)[1:5]

}
\references{
API docs at \url{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}
}
\seealso{
\code{rcax_nosa()}, \code{rcax_table_query()}, \code{rcax_filter()}, \code{rcax_key()}
}
