% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_rpers.R
\name{rcax_rpers}
\alias{rcax_rpers}
\title{Return RperS base table}
\usage{
rcax_rpers(
  tablename = "RperS",
  flist = NULL,
  qlist = NULL,
  cols = NULL,
  sortcols = c("broodyear", "popid"),
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)
}
\arguments{
\item{tablename}{(character) The CAX name for the table; ignored if \code{GETargs$table_id} passed in.}

\item{flist}{A filter for the query. See details and \code{rcax_filter()}}

\item{qlist}{Additional query params. See \code{rcax_table_query()} for details.}

\item{cols}{column names to return. Use cols=NULL if you want all columns}

\item{sortcols}{The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. \code{c("spawningyear", "popid", "MPG")}}

\item{type}{whether to return the table ("data.frame") or colnames ("colnames"). Default is to return the table as a data frame.}

\item{GETargs}{A list of arguments for the \code{rcax_GET()} call. These do not normally need to be specified by the user. \code{table_id} This is the CAX table id. It is looked up using \code{tablename} and \code{caxpops}. \code{recordloc} This is the name of the record we want in the list returned from the GET call. \code{key} is the API key. See \code{rcax_GET()} for API key details.}
}
\value{
data frame if \code{type="data.frame"} (default) and the colnames if \code{type="colnames"}
}
\description{
Returns the base HLI table which is missing the
additional metadata (such as NMFS_PopID and update dates) found in the XPort tables.

The query will download 1000 records by default; you can change
this by passing in \code{qlist=list(limit=1100)} for example. To make a
filtered query e.g. for one ESU or popid) pass in the \code{flist}
argument. Field names and values are case insensitive, e.g.
\code{POPID}, \code{popid} and \code{PopID} return the same result.
If your filter value is not in the table, the returned table
will be empty. Examples:
\itemize{
\item \code{flist=list(popid=7)} return values for popid 7.
\item \code{flist=list(esu_dps="Salmon, Chinook (Snake River spring/summer-run ESU)")} return values for one ESU. Use \code{rCAX:::caxesu} to see a list of ESU_DPS names.
}
}
\details{
The table_id is set automatically using a saved data frame from a \code{rcax_tables()} call. The table is saved in \code{R/sysdata.rda}.

If you want to see only the column names use
\verb{type="colnames}.
}
\examples{
# return popid of first record
# Note the part after $ is case sensitive
id <- rcax_rpers(qlist=list(limit=1))$popid

a <- rcax_rpers(
         flist=list(popid=id),
         cols=c("popid", "broodyear", "rpers")
         )
head(a)

# First 3 columns
rcax_rpers(qlist=list(limit=3))[,1:3]

# to print the first 5 column names
rcax_rpers(type="colnames")[1:5]

# to print the first 5 column names as they appear
# in Excel files downloaded from https://www.streamnet.org/data/hli/
tab <- rcax_rpers(qlist=list(limit=1))
colnames(tab)[1:5]

}
\references{
API docs at \url{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}
}
\seealso{
\code{rcax_rpers_xport()}, \code{rcax_table_query()}, \code{rcax_filter()}, \code{rcax_key()}
}
