% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_superpops.R
\name{rcax_superpops}
\alias{rcax_superpops}
\title{Get SuperPopulations table}
\usage{
rcax_superpops(
  popid = NULL,
  extra = NULL,
  table_id = NULL,
  key = NULL,
  parse = TRUE,
  cols = NULL,
  ...
)
}
\arguments{
\item{popid}{(character or number) A NMFS_POPID for a population}

\item{extra}{extra query params or column values. Enter as a list, e.g. list(pageno=7)}

\item{table_id}{not normally needed as this is set automatically to the table id for NOSA using the table saved from \code{rcax_table()}. But you can set it if needed.}

\item{key}{A CAX API token. See \code{\link{rcax_key}}.}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{cols}{column names to return. Use cols=NULL if you want all columns}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Returns SuperPopulations table

The SuperPopulations table_id is "009A08FE-6479-44FC-9B6F-01C55E2C8BA3" and
is set automatically using a saved data frame from a \code{rcax_tables()} call. The table is saved in \code{R/sysdata.rda}.

\code{rcax_superpops()} will download 1000 records.
}
\examples{
# to get all columns and print the column names
a <- rcax_superpops(cols = NULL)
colnames(a)

}
\references{
API docs at \url{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}

This function is modeled off the functions in \url{https://github.com/ropensci/rredlist}
}
