% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_table_query.R
\name{rcax_table_cols}
\alias{rcax_table_cols}
\title{Returns the colnames to use for tables}
\usage{
rcax_table_cols(hli, type = c("xport", "base"))
}
\arguments{
\item{hli}{The HLI short name: NOSA, SAR, PNI, RperS, JuvOut, or PreSmolt}

\item{type}{XPort table or base table}
}
\value{
A vector of column names
}
\description{
Returns the columns in the same order as the downloaded
Excel files have in https://www.streamnet.org/data/hli. Adds on
a few extra columns. The colnames are saved as internal data in
R/sysdata.rda and defined in inst/docs/create_sysdata.R
}
\details{
\itemize{
\item NOSA: Natural origin spawner abundance
\item SAR: SAR
\item PNI: PNI
\item RperS: Recruits per spawner
\item JuvOut: Short for JuvenileOutmigrants.
\item PreSmolt: Short for PresmoltAbundance.
}
}
\examples{
# Show the first 5 colnames
rcax_xport_cols("NOSA")[1:5]
}
