% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_GET.R
\name{rcax_table_name}
\alias{rcax_table_name}
\title{Return table name given the HLI short name}
\usage{
rcax_table_name(hli, type = c("xport", "base"))
}
\arguments{
\item{hli}{The HLI short name: NOSA, SAR, PNI, RperS, JuvOut, or PreSmolt}

\item{type}{XPort table or base table}
}
\description{
\itemize{
\item NOSA: Natural origin spawner abundance
\item SAR: SAR
\item PNI: PNI
\item RperS: Recruits per spawner
\item JuvOut: Short for JuvenileOutmigrants.
\item PreSmolt: Short for PresmoltAbundance.
}
}
\examples{
rcax_table_name("NOSA", type="xport")
}
