% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_table_query.R
\name{rcax_table_query}
\alias{rcax_table_query}
\title{Return tables by table name from Coordinated Assessments data eXchange}
\usage{
rcax_table_query(
  tablename = NULL,
  flist = NULL,
  qlist = NULL,
  cols = NULL,
  sortcols = NULL,
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)
}
\arguments{
\item{tablename}{The name of the table in the CAX API. See \code{rcax_tables()} for the names.}

\item{flist}{A filter for the query. See details and \code{rcax_filter()}}

\item{qlist}{Additional query params. See \code{rcax_table_query()} for details.}

\item{cols}{column names to return. Use cols=NULL if you want all columns. Names are case insensitive.}

\item{sortcols}{The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. \code{c("spawningyear", "popid", "MPG")}. Names are case insensitive.}

\item{type}{whether to return the table ("data.frame") or colnames with definitions ("colnames"). Default is to return the table as a data frame.}

\item{GETargs}{A list of arguments for the \code{rcax_GET()} call. These do not need to be specified by the user unless the user wants to change the default values. \code{table_id} This is the CAX table id. It is looked up using \code{tablename} and \code{rCAX:::caxpops} (an internal dataset). If \code{table_id} is passed in via \code{GETargs}, it wil override the default table id lookup. \code{recordloc} This is the name of the record we want in the list returned from the GET call. \code{key} is the API key.}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
data frame if \code{type="data.frame"} (default) and the colnames if \code{type="colnames"}
}
\description{
This is the base function for queries for most of the CAX tables.
Default queries will download 1000 records. You will
want to make a filtered query by passing in \code{flist} as a list with the column name values to filter on.
}
\details{
The required query parameters are table_id and XAPIKEY. These are set from the information in \code{GETargs}. table_id is normally looked up from \code{rcax_tables()} using \code{tablename} and \code{XAPIKEY} is looked up from \code{CAX_KEY} in the user environment (saved in \code{.Renviron} file. See \code{vignette("setup", package = "rCAX")} for how to set the key.  The \code{qlist} argument is any additional query parameters. The following are some of the available extra (meaning not required) query parameters.
\itemize{
\item page=num Integer of the page selected.
\item per_page=num  Integer for the number of records per page.
\item limit=num Maximum number of records to return. Default is 1000.
\item agency=character Agency that submitted the record. You can designate one or more (comma-separated) agency acronyms: IDFG, WDFW, ODFW, CRITFC, CCT, NOAA.
\item updated_since	Date since epoch in seconds: 2012-04-24 15:05:22 -0400 = 1335294322.
\item popid Population id in NOSA tables.
}

Filtering, i.e. subsetting, the query is specified using the \code{flist} argument.
This is a list with the column name and value. See \code{rcax_filter()} for the filtering code which writes the filter query parameter that is added to the GET query. See \code{rcax_filter()} for examples of how one passes in \code{flist}.
}
\examples{
a <- rcax_table_query(tablename="NOSA", qlist = list(limit=5))
a[, c("popid", "spawningyear", "nosaij", "nosaej")]

}
\references{
API docs at \url{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}
}
\seealso{
\code{rcax_GET()}, \code{rcax_hli()}, \code{rcax_escdata()}, \code{rcax_superpops()}
}
