% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_table_query.R
\name{rcax_table_sortcols}
\alias{rcax_table_sortcols}
\title{Returns the default sort columns for tables}
\usage{
rcax_table_sortcols(hli, type = c("xport", "base"))
}
\arguments{
\item{hli}{The HLI short name: NOSA, SAR, PNI, RperS, JuvOut, or PreSmolt}

\item{type}{XPort table or base table}
}
\value{
A vector of column names for sorting
}
\description{
\itemize{
\item NOSA = c("spawningyear", "popid")
\item SAR = c("outmigrationyear", "popid")
\item PNI = c("spawningyear", "popid")
\item JuvOut = c("outmigrationyear", "popid")
\item PreSmolt = c("surveyyear", "popid")
\item RperS = c("broodyear", "popid")
}
}
\examples{
rcax_table_sortcols("NOSA")
}
