% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_tables.R
\name{rcax_tables}
\alias{rcax_tables}
\title{Get the list of CAX tables with ids}
\usage{
rcax_tables(
  key = NULL,
  parse = TRUE,
  cols = c("name", "id", "description"),
  ...
)
}
\arguments{
\item{key}{A CAX API key. See \code{vignette("setup", package = "rCAX")}.}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{cols}{the columns to return. default is name, description and id. use cols=NULL to get all the columns}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Get the list of CAX tables with ids
}
\examples{
a <- rcax_tables(cols=NULL)
colnames(a)

head(a[,1:2])

}
\references{
API docs at \url{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}
}
