% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_termsofuse.R
\name{rcax_termsofuse}
\alias{rcax_termsofuse}
\title{Get CAX HLI data Terms of Use}
\usage{
rcax_termsofuse()
}
\value{
Text of the terms of us
}
\description{
Get CAX HLI data Terms of Use
}
\examples{
rcax_termsofuse()


}
