% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_GET.R
\name{rcax_ua}
\alias{rcax_ua}
\title{Make the user agent string}
\usage{
rcax_ua()
}
\description{
Make the user agent string
}
\examples{
rcax_ua()
}
\references{
This function is modeled off the functions in \url{https://github.com/ropensci/rredlist}
}
