% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_key.R
\name{rcax_use_cax}
\alias{rcax_use_cax}
\title{Print message re how to get CAX API key}
\usage{
rcax_use_cax()
}
\description{
Provides instruction on how to store the key.
}
\details{
Contact Streamnet to request a pull key. https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/
}
