% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_version.R
\name{rcax_version}
\alias{rcax_version}
\title{Look up the rCAX versions}
\usage{
rcax_version()
}
\value{
returns the GitHub version invisibly
}
\description{
Get the version installed and the version on GitHub.
}
\examples{
rcax_version()

}
