% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_citation.R
\name{rl_citation}
\alias{rl_citation}
\title{Get the citation CAX version}
\usage{
rl_citation(...)
}
\arguments{
\item{...}{Curl options passed to [crul::verb-GET]}
}
\value{
API citation as character string
}
\description{
Get the citation CAX version
}
\examples{
\dontrun{
rcax_citation()
}
}
