% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax-package.R
\docType{package}
\name{rCAX-package}
\alias{rCAX-package}
\alias{rCAX}
\title{rCAX}
\description{
This package is an R client for the StreamNet Coordinated Assessments HLI REST API.

Documentation: https://nwfsc-cb.github.io/rCAX/
}
\section{Authentication}{

The rCAX R package includes a pull key so that you can query the CAX database. Should you
need to use your own key see the vignette on the API functions.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nwfsc-cb/rCAX}
  \item Report bugs at \url{https://github.com/nwfsc-cb/rCAX/issues}
}

}
\author{
\strong{Maintainer}: Eli Holmes \email{eli.holmes@noaa.gov}

Authors:
\itemize{
  \item Mari Williams \email{mari.williams@noaa.gov}
}

Other contributors:
\itemize{
  \item Katie Barnas \email{katie.barnas@noaa.gov} [contributor]
}

}
\keyword{package}
