% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_filter.R
\name{rcax_filter}
\alias{rcax_filter}
\title{Create filter to add to GET call}
\usage{
rcax_filter(flist)
}
\arguments{
\item{flist}{list of property (column) names, values in this format \code{list(colname=value)}}

\item{type}{(optional) a list with type for property names. If not passed in, the type will be inferred from the value in \code{flist}.}
}
\description{
Create filter to add to GET call
}
\details{
\itemize{
\item Any spaces in the values in \code{flist} will be replaced with \verb{\%20}. This is done automatically but can also be done by the user.
\item \code{type} The type of the value, for example "string" will be inferred from \code{class(flist$value)} but if you run into problems, you can pass in type for that property. It is not necessary to include all property names in \code{flist}. Only properties for which \code{class(flist$value)} is not producing the correct type.
\item The property names in \code{flist} are the column names in a table being returned by \code{rcax_GET()}.
}
}
\examples{
\dontrun{
rcax_GET("ca/tables")
}
}
\author{
EE Holmes
}
