% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_GET.R
\name{rcax_GET}
\alias{rcax_GET}
\title{Make a GET call to API}
\usage{
rcax_GET(path, key = NULL, query = NULL, ...)
}
\arguments{
\item{path}{what to add after the base api path}

\item{key}{the api key}

\item{...}{extra info}
}
\description{
Make a GET call to API
}
\examples{
\dontrun{
rcax_GET("ca/tables")
}
}
\references{
This function is modeled off the functions in \url{https://github.com/ropensci/rredlist}
}
