% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_GET.R
\name{rcax_GET}
\alias{rcax_GET}
\title{Make a GET call to API}
\usage{
rcax_GET(path, key = NULL, query = NULL, ...)
}
\arguments{
\item{path}{what to add after the base api path}

\item{key}{A CAX API key. See \code{vignette("setup", package = "rCAX")}.}

\item{query}{a list of query parameters with their values, e.g. \code{list(param=value)}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Make a GET call to API
}
\details{
You will need an API key to make GET queries. Contact Streamnet to request a pull key or guest key. https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/

After getting your key set it as CAX_KEY in your \code{.Renviron} file by pasting this text into that file. \code{CAX_KEY='youractualkeynotthisstring'}

To find (and create if necessary) your \code{.Renviron} file, use \code{usethis::edit_r_environ()}
then open that file and paste in the key text above.
After you edit the file, you will need to restart R. If you are in RStudio use Session > Restart R.
}
\examples{
\dontrun{
rcax_GET("ca/tables")
}
}
\references{
API docs at \url{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}

This function is modeled off \code{rl_GET()} in \url{https://github.com/ropensci/rredlist}
}
