% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_nosa.R
\name{rcax_nosa}
\alias{rcax_nosa}
\title{Return NOSA table}
\usage{
rcax_nosa(
  tablename = "NOSA",
  flist = NULL,
  qlist = NULL,
  cols = c("recoverydomain", "esu_dps", "commonpopname", "commonname", "run", "popid",
    "majorpopgroup", "spawningyear", "tsaij", "tsaej", "nosaij", "nosaej"),
  sortcols = c("spawningyear", "popid"),
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)
}
\arguments{
\item{tablename}{(character) The CAX name for the table; ignored if \code{GETargs$table_id} passed in.}

\item{flist}{A filter for the query. See details and \code{rcax_filter()}}

\item{qlist}{Additional query params. See \code{rcax_table_query()} for details.}

\item{cols}{column names to return. Use cols=NULL if you want all columns}

\item{sortcols}{The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. \code{c("spawningyear", "popid", "MPG")}}

\item{type}{whether to return the table ("data.frame") or colnames ("colnames"). Default is to return the table as a data frame.}

\item{GETargs}{A list of arguments for the \code{rcax_GET()} call. These do not normally need to be specified by the user. \code{table_id} This is the CAX table id. It is looked up using \code{tablename} and \code{caxpops}. \code{recordloc} This is the name of the record we want in the list returned from the GET call. \code{key} is the API key. See \code{rcax_GET()} for API key details.}
}
\value{
data frame if \code{type="data.frame"} (default) and the colnames if \code{type="colnames"}
}
\description{
Returns NOSA table sorted by popid and spawningyear. The default
columns to return are "recoverydomain", "esu_dps", "commonpopname",
"commonname", "run", "popid", "majorpopgroup", "spawningyear", "tsaij", "tsaej",
"nosaij", "nosaej"

\code{rcax_nosa()} will download 1000 records from the NOSA table. You will
probably want to make a filtered query by passing in value to filter on
via \code{flist}. See \code{rcax_filter()}. Examples for the NOSA table:
\itemize{
\item \verb{rcax_nosa(flist=list(popid=7)} return values for popid 7.
\item \verb{rcax_nosa(flist=list(esu_dps="Salmon, Chinook (Snake River spring/summer-run ESU)")} return values for one ESU. Use \code{rCAX:::caxesu} to see a list of ESU_DPS names.
}
}
\details{
The NOSA table_id is "4EF09E86-2AA8-4C98-A983-A272C2C2C7E3" and
is set automatically using a saved data frame from a \code{rcax_tables()} call. The table is saved in \code{R/sysdata.rda}.
}
\examples{
a <- rcax_nosa(flist=list(popid=7))
a[, c("popid", "spawningyear", "nosaij", "nosaej")] |> head()

# to print the first 5 column names
rcax_nosa(type="colnames")[1:5]

}
\references{
API docs at \url{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}
}
\seealso{
\code{rcax_nosa_xport()}, \code{rcax_table_query()}, \code{rcax_filter()}, \code{rcax_key()}
}
