% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_nosa.R
\name{rcax_nosa}
\alias{rcax_nosa}
\alias{rcax_nosa_}
\title{Get nosa table}
\usage{
rcax_nosa(table_id, key = NULL, parse = TRUE, ...)

rcax_nosa_(table_id, key = NULL, ...)
}
\arguments{
\item{table_id}{table id eg 4EF09E86-2AA8-4C98-A983-A272C2C2C7E3}

\item{key}{A CAX API token. See \code{\link{rcax_use_iucn}}.}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
A list, with the data in the \code{result} slot, unless using
a function with a trailing underscore, in which case json as character
string is returned.
}
\description{
Get nosa table
}
\examples{
\dontrun{
rcax_nosa("4EF09E86-2AA8-4C98-A983-A272C2C2C7E3")
rcax_nosa_()
}
}
\references{
API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/
}
