% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_GET.R
\name{rcax_parse}
\alias{rcax_parse}
\title{Parse results from API call}
\usage{
rcax_parse(x, parse)
}
\arguments{
\item{x}{what needs to be parsed}

\item{parse}{logical true/false}
}
\description{
Parse results from API call
}
