% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_superpops.R
\name{rcax_superpops}
\alias{rcax_superpops}
\title{Get SuperPopulations table}
\usage{
rcax_superpops(
  tablename = "SuperPopulations",
  flist = NULL,
  qlist = NULL,
  cols = NULL,
  sortcols = c("popid"),
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)
}
\arguments{
\item{tablename}{The name of the table in the CAX API. See \code{vignette("CAX_Tables", package = "rCAX")}}

\item{flist}{A filter for the query. See details and \code{rcax_filter()}}

\item{qlist}{Additional query params. See \code{rcax_table_query()} for details.}

\item{cols}{column names to return. Use cols=NULL if you want all columns. Names are case insensitive.}

\item{sortcols}{The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. \code{c("spawningyear", "popid", "MPG")}. Names are case insensitive.}

\item{type}{whether to return the table ("data.frame") or colnames with definitions ("colnames"). Default is to return the table as a data frame.}

\item{GETargs}{A list of arguments for the \code{rcax_GET()} call. These do not need to be specified by the user unless the user wants to change the default values. \code{table_id} This is the CAX table id. It is looked up using \code{tablename} and \code{rCAX:::caxpops} (an internal dataset). If \code{table_id} is passed in via \code{GETargs}, it wil override the default table id lookup. \code{recordloc} This is the name of the record we want in the list returned from the GET call. \code{key} is the API key.}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
data frame if \code{type="data.frame"} (default) and the colnames if \code{type="colnames"}
}
\description{
Returns SuperPopulations table

The SuperPopulations table_id is set automatically using a saved data frame from a \code{rcax_tables()} call. The table is saved in \code{R/sysdata.rda}.

\code{rcax_superpops()} will download 1000 records. Pass in \code{flist} to filter.
}
\examples{
# print the first columns and definitions
a <- rcax_superpops(type="colnames")
paste(a$name, a$definition, sep=": ")[1:5]

# get one record
rcax_superpops(qlist=list(limit=1))

}
\references{
API docs at \url{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}
}
\seealso{
\code{rcax_table_query()}, \code{rcax_filter()}
}
