% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_tables.R
\name{rcax_tables}
\alias{rcax_tables}
\alias{rcax_tables_}
\title{Get the list of CAX tables with ids}
\usage{
rcax_tables(
  key = NULL,
  parse = TRUE,
  cols = c("name", "id", "description"),
  ...
)

rcax_tables_(key = NULL, ...)
}
\arguments{
\item{key}{A CAX API token. See \code{\link{rcax_use_iucn}}.}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). Default: \code{TRUE}}

\item{cols}{the columns to return. default is name, description and id}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\value{
A list, with the data in the \code{result} slot, unless using
a function with a trailing underscore, in which case json as character
string is returned.
}
\description{
Get the list of CAX tables with ids
}
\examples{
head(rcax_tables()[,1:2])

}
\references{
API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/
}
