% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_tables.R
\name{rcax_tables}
\alias{rcax_tables}
\title{Get the list of CAX tables with ids}
\usage{
rcax_tables(
  cols = c("name", "id", "description"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)
}
\arguments{
\item{cols}{the columns to return. default is name, description and id. use cols=NULL to get all the columns}

\item{GETargs}{A list of arguments for the \code{rcax_GET()} call. These do not need to be specified by the user unless the user wants to change the default values. \code{table_id} This is the CAX table id. It is looked up using \code{tablename} and \code{rCAX:::caxpops} (an internal dataset). If \code{table_id} is passed in via \code{GETargs}, it wil override the default table id lookup. \code{recordloc} This is the name of the record we want in the list returned from the GET call. \code{key} is the API key.}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Get the list of CAX tables with ids
}
\examples{
a <- rcax_tables(cols=NULL)
colnames(a)

head(a[,1:2])

}
\references{
API docs at \url{https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/}
}
