% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcax_version.R
\name{rcax_version}
\alias{rcax_version}
\title{Get the CAX API version}
\usage{
rcax_version(...)
}
\arguments{
\item{...}{Curl options passed to [crul::verb-GET]}
}
\value{
API version as character string
}
\description{
Get the CAX API version
}
\examples{
\dontrun{
rcax_version()
}
}
